+++ src/Plugin/src/Export/IgcExport/src/IgcExportSettings.cpp
@@ -37,10 +37,6 @@ namespace
     // Keys
     constexpr char PilotNameKey[] = "PilotName";
     constexpr char CoPilotNameKey[] = "CoPilotName";
-    constexpr char ConvertAltitudeKey[] = "ConvertAltitude";
-
-    // Defaults
-    constexpr bool DefaultConvertAltitude {true};
 }
 
 class IgcExportSettingsPrivate
@@ -48,13 +44,11 @@ class IgcExportSettingsPrivate
 public:
     IgcExportSettingsPrivate()
         : pilotName(DefaultPilotName),
-          coPilotName(DefaultCoPilotName),
-          convertAltitude(::DefaultConvertAltitude)
+          coPilotName(DefaultCoPilotName)
     {}
 
     QString pilotName;
     QString coPilotName;
-    bool convertAltitude;
 
     static inline const QString DefaultPilotName {System::getUsername()};
     static inline const QString DefaultCoPilotName {QString()};
@@ -104,19 +98,6 @@ void IgcExportSettings::setCoPilotName(const QString &coPilotName) noexcept
     }
 }
 
-bool IgcExportSettings::isConvertAltitudeEnabled() const noexcept
-{
-    return d->convertAltitude;
-}
-
-void IgcExportSettings::setConvertAltitudeEnabled(bool enable) noexcept
-{
-    if (d->convertAltitude != enable) {
-        d->convertAltitude = enable;
-        emit extendedSettingsChanged();
-    }
-}
-
 // PROTECTED
 
 void IgcExportSettings::addSettingsExtn(Settings::KeyValues &keyValues) const noexcept
@@ -129,10 +110,6 @@ void IgcExportSettings::addSettingsExtn(Settings::KeyValues &keyValues) const no
 
     keyValue.first = ::CoPilotNameKey;
     keyValue.second = d->coPilotName;
-    keyValues.push_back(keyValue);
-
-    keyValue.first = ::ConvertAltitudeKey;
-    keyValue.second = d->convertAltitude;
     keyValues.push_back(keyValue);
 }
 
@@ -147,17 +124,12 @@ void IgcExportSettings::addKeysWithDefaultsExtn(Settings::KeysWithDefaults &keys
     keyValue.first = ::CoPilotNameKey;
     keyValue.second = IgcExportSettingsPrivate::DefaultCoPilotName;
     keysWithDefaults.push_back(keyValue);
-
-    keyValue.first = ::ConvertAltitudeKey;
-    keyValue.second = ::DefaultConvertAltitude;
-    keysWithDefaults.push_back(keyValue);
 }
 
 void IgcExportSettings::restoreSettingsExtn(const Settings::ValuesByKey &valuesByKey) noexcept
 {
     d->pilotName = valuesByKey.at(::PilotNameKey).value<QString>();
     d->coPilotName = valuesByKey.at(::CoPilotNameKey).value<QString>();
-    d->convertAltitude = valuesByKey.at(::ConvertAltitudeKey).toBool();
 
     emit extendedSettingsChanged();
 }
@@ -166,7 +138,6 @@ void IgcExportSettings::restoreDefaultsExtn() noexcept
 {
     d->pilotName = IgcExportSettingsPrivate::DefaultPilotName;
     d->coPilotName = IgcExportSettingsPrivate::DefaultCoPilotName;
-    d->convertAltitude = ::DefaultConvertAltitude;
 
     emit extendedSettingsChanged();
 }